[Remote-Mediator] Change PagingSource (RemoteMediator)

+++ domain/src/main/java/com/nanamare/domain/model/GenreListModel.kt
@@ -2,4 +2,4 @@ package com.nanamare.domain.model
 
 data class GenreListModel(val genres: List<GenreModel>)
 
-data class GenreModel(val id: Int, val name: String)
+data class GenreModel(val id: Long, val name: String)
